Adding timeout option for validating the camera is alive.
Middlebox doesnt know the status of cameras. Sometimes cameras fail network
connectivity to the middlebox. While copying files to the webserver, relay
agent validates the size of each file chunk. When the size of file less than
min size for a given elapsed time. middlebox will mark the camera to
disconnected state.
Its responsibility of user to remove and readd the camera again for getting the
streaming back.

Signed-off-by: Sugesh Chandran <sugesh.chandran@intel.com>

Add the timer library for the timeout support

Signed-off-by: Sugesh Chandran <sugesh.chandran@intel.com>