nvmidbox-ws : Middlebox switch page to run on http than https.
Adding liveview support need the switch page to be http as vlc generate
http stream url. So its very difficult to embed it into https web page.

Changing the webpage to http on port 9000 than 8080. The websockets are
still secured(wss) and they are served on same port 8080.
The websocket connections are made from external entities such as neoview
webserver. So keep it secure under wss.

Signed-off-by: Sugesh Chandran <sugesh.chandran@intel.com>