midbox-conf : Adding a new function to update liveurl.
Middlebox cannot track connectivity to the camera. The issue with loosing
connectivity to the camera is, liveview is getting stopped and cannot restart.

The fix for that issue is to terminate and restart liveview on fixed time
interval. This will ensure intermittent camera issues doesnot effect the live
preview implementation.

Signed-off-by: Sugesh Chandran <sugesh.chandran@intel.com>