Update locking lib to keep track of threads that using it.
The changes made into the lock libraries to keep track of threads when its
acquiring a lock. It also keep track of time at which the lock is being
acquired by the thread.

This changes will be used in the following commits, In other words the
thread and time information will be used in watchdog thread. Following commits
will add support for the watchdog.

Signed-off-by: sugu <sugeshchandran@gmail.com>